We have created the GameController class that will replace RendereTest and will be in charge of managing both the model and the controller.
We have created the GameKeyListener class, which will be in charge of
managing the keystrokes in order to interact with the window.

We have fixed the bug that added score when you hit a box even if you
don't move.

We have fixed the bug that caused difficulties when moving the character